+++ src/plugins/e-acsl/tests/runtime/oracle/gen_typedef.c
@@ -7,8 +7,8 @@ int main(void)
   __e_acsl_memory_init((int *)0,(char ***)0,(size_t)8);
   uint8 x = (unsigned char)0;
   /*@ assert x ≡ 0; */
-  __e_acsl_assert(x == 0,(char *)"Assertion",(char *)"main",(char *)"x == 0",
-                  9);
+  __e_acsl_assert((int)x == 0,(char *)"Assertion",(char *)"main",
+                  (char *)"x == 0",9);
   __retres = 0;
   return __retres;
 }
